requirements.txt
fastapi==0.103.1
uvicorn==0.24.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.8
pydantic==2.4.2
jinja2==3.1.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
python-jose==3.3.0